move to Cms Editor Descriptors
