--- src/Foundation.Cms/Extensions/IDisableOPE.cs
@@ -1,6 +1,0 @@
-﻿namespace Foundation.Cms.Extensions
-{
-    public interface IDisableOPE
-    {
-    }
-}
