+++ src/Foundation.Commerce/Models/Pages/OrganizationPage.cs
@@ -1,6 +1,6 @@
 using EPiServer.DataAbstraction;
 using EPiServer.DataAnnotations;
-using Foundation.Cms.Extensions;
+using Foundation.Cms.EditorDescriptors;
 using Foundation.Cms.Pages;
 
 namespace Foundation.Commerce.Models.Pages
